+++ src/Sulu/Bundle/MediaBundle/Controller/MediaController.php
@@ -117,7 +117,8 @@ class MediaController extends AbstractMediaController implements
         $locale = $this->getRequestParameter($request, 'locale', true);
         $fieldDescriptors = $this->getFieldDescriptors($locale, false);
         $ids = array_filter(explode(',', $request->get('ids')));
-        $listBuilder = $this->getListBuilder($request, $fieldDescriptors, $ids);
+        $types = array_filter(explode(',', $request->get('types')));
+        $listBuilder = $this->getListBuilder($request, $fieldDescriptors, $ids, $types);
         $listResponse = $listBuilder->execute();
         $count = $listBuilder->count();
 
@@ -170,10 +171,12 @@ class MediaController extends AbstractMediaController implements
      *
      * @param Request $request
      * @param FieldDescriptorInterface[] $fieldDescriptors
+     * @param array $ids
+     * @param array $types
      *
      * @return DoctrineListBuilder
      */
-    private function getListBuilder(Request $request, array $fieldDescriptors, $ids)
+    private function getListBuilder(Request $request, array $fieldDescriptors, $ids, $types)
     {
         $restHelper = $this->get('sulu_core.doctrine_rest_helper');
         $factory = $this->get('sulu_core.doctrine_list_builder_factory');
@@ -209,6 +212,11 @@ class MediaController extends AbstractMediaController implements
             }
 
             $listBuilder->in($fieldDescriptors['id'], $ids);
+        }
+
+        // set the types
+        if (count($types)) {
+            $listBuilder->in($fieldDescriptors['type'], $types);
         }
 
         // field which will be needed afterwards to generate route